1853 Cholera Outbreak Analysis
Stockholm's Cholera Outbreak as archived by St. Katarina's Church in 1853 was another interesting topic for my Data Analysis Project 5/12 - powered by Quantum Analytics NG.

History:
In the 19th century, the Church was in charge of the world's population up until 1870. St. Catherine's (Katarina) Catholic Church in Stockholm Sweden, had the privilege of a handwritten record of the second most fierce pandemic at the time, after the 1834 pandemic of the same Stockholm that took over 3,000 lives - which were approximately 3% of the city's population. Both the 1834 and 1853 outbreaks crossed the death margin of 3,000 and none other again reached the death mark out of the 3 Cholera pandemics. However, our dataset has only 248 records.

Insights:
The outbreak had a death speed of 4 daily from August to November. September had the highest recorded at 183 deaths; 6 deaths on an average day and also recorded the highest number of deaths in a single day (15). The pandemic ripped all age groups but predominantly middle-aged adults (35 to 55) and babies less than 5 years of age. More female deaths occurred with a slight difference of 12 between the gender