Historical Stock Prices of 10 Popular Companies
Here is one of my full-stacked Data Analytics projects, Historical Stock Prices of 10 Popular Companies from 2015 to 2021, which I have been working on this summer as facilitated by Quantum Analytics NG. This is project 1/12, but there have been a lot of picked-up skills and techniques, especially in the exploratory analysis part.

Insights:
Overall, the 10 companies produced a massive figure in terms of volume of trading shares, prices, and their trends year-on-year.

Amazon and Google were the most valued shares in the market at the end of the period. In 6 years, the values of their shares grew from less than a thousand dollars to about $3,000. However, as the values of these shares were increasing tremendously so was the trading volume of shares declining in its fashion. This was evident with Apple hitting almost 140 million average volume shares a year at just $61.

Only Tesla and Microsoft maintained an increasing change in their share prices and were the only companies that showed no decline in 2018. Tesla has never shown a decline on average and has the best growth in terms of open and closed prices, followed by Zoom. In fact, Zoom topped the chart in 2020 when it made a 0.42% increase on average. The Covid year was the Zoom year!